backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
youtube-transcript-api==0.6.1
google-api-python-client==2.108.0
pytest==7.4.3
httpx==0.25.2